updated inputs
