+++ package.json
@@ -42,15 +42,15 @@
     "@radio-retail/babel-preset": "1.0.0",
     "@radio-retail/eslint-config": "1.0.0",
     "@rollup/plugin-json": "^4.1.0",
-    "@storybook/addon-a11y": "^5.3.19",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addon-docs": "^5.3.19",
+    "@storybook/addon-a11y": "^5.3.21",
+    "@storybook/addon-actions": "^5.3.21",
+    "@storybook/addon-docs": "^5.3.21",
     "@storybook/addon-knobs": "5.3.19",
-    "@storybook/addon-links": "^5.3.19",
+    "@storybook/addon-links": "^5.3.21",
     "@storybook/addon-storyshots": "5.3.19",
     "@storybook/addon-storyshots-puppeteer": "5.3.19",
-    "@storybook/addon-storysource": "^5.3.19",
-    "@storybook/addon-viewport": "^5.3.19",
+    "@storybook/addon-storysource": "^5.3.21",
+    "@storybook/addon-viewport": "^5.3.21",
     "@storybook/codemod": "5.3.19",
     "@storybook/preset-create-react-app": "^3.1.4",
     "@storybook/react": "5.3.19",
@@ -68,9 +68,9 @@
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-react": "7.20.5",
     "eslint-plugin-react-hooks": "4.0.8",
-    "husky": "^4.2.5",
+    "husky": "^4.3.0",
     "jest": "24.9.0",
-    "lint-staged": "^10.2.10",
+    "lint-staged": "^10.4.0",
     "markdown-loader-jest": "0.1.1",
     "preact-render-to-json": "3.6.6",
     "prettier": "1.19.1",
@@ -78,8 +78,8 @@
     "prop-types": "15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-hook-form": "^6.8.3",
-    "react-scripts": "^3.4.1",
+    "react-hook-form": "^6.8.6",
+    "react-scripts": "^3.4.3",
     "react-test-renderer": "16.13.1",
     "rollup": "2.26.3",
     "rollup-plugin-babel": "4.4.0",
@@ -87,12 +87,12 @@
     "rollup-plugin-css-porter": "1.0.2",
     "rollup-plugin-filesize": "6.2.1",
     "rollup-plugin-node-resolve": "5.2.0",
-    "rollup-plugin-visualizer": "^4.0.4",
-    "snyk": "^1.361.3",
+    "rollup-plugin-visualizer": "^4.1.1",
+    "snyk": "^1.401.0",
     "storybook-addon-jsx": "7.1.13",
     "storybook-readme": "5.0.8",
     "style-loader": "1.2.1",
-    "styled-components": "^5.1.1",
+    "styled-components": "^5.2.0",
     "stylelint": "13.6.1",
     "stylelint-config-recommended": "3.0.0",
     "stylelint-config-styled-components": "0.1.1",
