+++ src/components/input/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+import { inputStyles } from "../input-styles";
+
+export const StyledInput = styled.input`
+  height: 35px;
+  ${inputStyles}
+`;
