+++ src/components/select/index.js
@@ -1,7 +1,7 @@
 import React from "react";
-import InputField from "../input-field";
+import Input from "../input";
 import withAutocomplete from "../with-auto-complete";
 
-const Select = props => <InputField {...props} fullWidth />;
+const Select = props => <Input {...props} fullWidth noRef />;
 
 export default withAutocomplete(Select);
