+++ src/components/tel-input/styles.js
@@ -1,39 +1,14 @@
 import styled from "styled-components";
-import { colors, fontSizes, space, radii } from "../theme";
-import InputField from "../input-field";
+import { inputStyles } from "../input-styles";
 
-export const StyledInput = styled.input`
+export const StyledCodeInput = styled.input`
   width: 7em;
   height: 35px;
-  outline: none;
-  margin-top: 1.3em;
-  position: absolute;
-  color: ${colors.gray.xxdark};
-  border-radius: ${radii.small};
-  font-size: ${fontSizes.xsmall};
-  padding: ${space[1]} ${space[2]};
-  background-color: ${colors.white};
-  border: 1px solid ${colors.gray.xxlight};
-
-  ::placeholder {
-    color: ${colors.gray.dark};
-  }
-
-  &:disabled {
-    background-color: ${colors.gray.xxlight};
-  }
-
-  ::-webkit-inner-spin-button,
-  ::-webkit-outer-spin-button {
-    /* stylelint-disable-next-line property-no-vendor-prefix */
-    -webkit-appearance: none;
-    margin: 0;
-  }
+  ${inputStyles}
 `;
 
-export const StyledInputField = styled(InputField).attrs(() => ({
-  inputStyle: {
-    alignSelf: "flex-end",
-    width: `calc(100% - 7em - ${space[4]})`
-  }
-}))``;
+export const StyledNumberInput = styled.input`
+  width: 100%;
+  height: 35px;
+  ${inputStyles}
+`;
