+++ src/components/text-area/styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { inputStyles } from "components/inputs/styles";
+
+export const StyledTextArea = styled.textarea`
+  resize: none;
+  min-height: 10em;
+  font-family: inherit;
+  ${inputStyles}
+`;
